django_server/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.5
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
idna==3.7
jwcrypto==1.5.6
mysqlclient==2.2.4
oauthlib==3.2.2
pycparser==2.22
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
